Add files via upload
Slakteri data was taken from Data Norge. This is a data set that shows deliveries of slaughter and wool delivered from agricultural enterprises to slaughterhouses. For the individual enterprise, the organization number, name, municipality number and kilos delivered to the slaughterhouse during the calendar year are shown. The data does not include rent cuts. A new data set is published annually. This happens approx. 15 March, the year after the animals have been delivered to the slaughterhouse.